portional-web/requirements.txt
django
psycopg2
djangorestframework
markdown
django-filter
django-model-utils
